+++ frontend/src/app/core/models/warehouse-item.ts
